Fix storage of account value
The missing key for the call to set results in each character of the accountID being stored against numerical keys in the localstorage.

This hasn't yet manifested as a problem because the `*` route also calls to set this value before redirecting. If you remove that logic in the route definitions this provider would not work.